+++ R/validate.R
@@ -33,7 +33,7 @@ dm_validate <- function(x) {
 
   check_df_structure(def, boilerplate, "dm definition")
 
-  if (!all(map_lgl(def$data, ~ inherits(., "data.frame") || inherits(., "tbl_dbi")))) {
+  if (!all(map_lgl(def$data, ~ inherits(., "data.frame") || inherits(., "tbl_sql")))) {
     abort_dm_invalid(
       "Not all entries in `def$data` are of class `data.frame` or `tbl_dbi`. Check `dm_get_tables()`."
     )
@@ -64,7 +64,11 @@ dm_validate <- function(x) {
     select(table = ref_table, column = ref_column) %>%
     check_colnames(dm_col_names, "Parent key")
 
-  stopifnot(lengths(def$pks) %in% 0:1)
+  if (!all(map_int(def$pks, vctrs::vec_size) %in% 0:1)) {
+    abort_dm_invalid(
+      "Not all tables have maximally 1 primary key."
+    )
+  }
 
   pks <-
     def %>%
@@ -81,6 +85,14 @@ dm_validate <- function(x) {
       "Not all entries in `def$zoom` are of class `data.frame`, `tbl_dbi` or `NULL`."
     )
   }
+
+  uks <- def %>%
+    select(table, uks) %>%
+    unnest_list_of_df("uks")
+
+  uks %>%
+    unnest_col("column", character()) %>%
+    check_colnames(dm_col_names, "UK")
 
   check_no_nulls(def)
 
@@ -107,9 +119,9 @@ debug_dm_validate <- function(dm) {
   dm
 }
 
-check_dm <- function(dm) {
-  if (!is_dm(dm)) {
-    abort_is_not_dm(class(dm))
+check_dm <- function(x) {
+  if (!is_dm(x)) {
+    abort_is_not_dm(class(x))
   }
 }