+++ tests/testthat/helper-expectations.R
@@ -2,20 +2,28 @@ expect_identical_graph <- function(g1, g2) {
   expect_true(igraph::identical_graphs(g1, g2))
 }
 
-expect_equivalent_dm <- function(object, expected, sort = FALSE, ..., sort_tables = sort, sort_columns = sort, sort_keys = sort, ignore_on_delete = FALSE) {
+expect_equivalent_dm <- function(object, expected, sort = FALSE, ..., sort_tables = sort, sort_columns = sort, sort_keys = sort, ignore_on_delete = FALSE, ignore_autoincrement = FALSE) {
   tables1 <- dm_get_tables_impl(object) %>% map(collect)
   tables2 <- dm_get_tables_impl(expected) %>% map(collect)
 
   expect_equivalent_tbl_lists(tables1, tables2, sort_tables = sort_tables, sort_columns = sort_columns)
 
   if (sort_keys) {
-    expect_equivalent_tbl(dm_get_all_pks_impl(object), dm_get_all_pks_impl(expected))
+    if (ignore_autoincrement) {
+      expect_equivalent_tbl(dm_get_all_pks_impl(object) %>% select(-autoincrement), dm_get_all_pks_impl(expected) %>% select(-autoincrement))
+    } else {
+      expect_equivalent_tbl(dm_get_all_pks_impl(object), dm_get_all_pks_impl(expected))
+    }
     expect_equivalent_tbl(
       dm_get_all_fks_impl(object, ignore_on_delete = ignore_on_delete),
       dm_get_all_fks_impl(expected, ignore_on_delete = ignore_on_delete)
     )
   } else {
-    expect_equal(dm_get_all_pks_impl(object), dm_get_all_pks_impl(expected))
+    if (ignore_autoincrement) {
+      expect_equivalent_tbl(dm_get_all_pks_impl(object) %>% select(-autoincrement), dm_get_all_pks_impl(expected) %>% select(-autoincrement))
+    } else {
+      expect_equivalent_tbl(dm_get_all_pks_impl(object), dm_get_all_pks_impl(expected))
+    }
     expect_equal(
       dm_get_all_fks_impl(object, ignore_on_delete = ignore_on_delete),
       dm_get_all_fks_impl(expected, ignore_on_delete = ignore_on_delete)