+++ tests/testthat/test-deconstruct.R
@@ -117,6 +117,8 @@ test_that("`dm()` and `new_dm()` can handle a list of `dm_keyed_tbl` objects", {
 })
 
 test_that("`dm()` and `new_dm()` can handle a mix of tables and `dm_keyed_tbl` objects", {
+  skip_if_not_installed("nycflights13")
+
   dm <- dm_nycflights13(cycle = TRUE)
 
   y1 <- keyed_tbl_impl(dm, "weather") %>%
@@ -190,11 +192,11 @@ test_that("joins without child PK", {
 
   expect_snapshot({
     keyed_build_join_spec(x, y) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(x, y)) %>%
       dm_paste(options = c("select", "keys"))
     keyed_build_join_spec(y, x) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(y, x)) %>%
       dm_paste(options = c("select", "keys"))
   })
@@ -214,11 +216,11 @@ test_that("joins with other child PK", {
 
   expect_snapshot({
     keyed_build_join_spec(x, y) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(x, y)) %>%
       dm_paste(options = c("select", "keys"))
     keyed_build_join_spec(y, x) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(y, x)) %>%
       dm_paste(options = c("select", "keys"))
   })
@@ -238,11 +240,11 @@ test_that("joins with other child PK and name conflict", {
 
   expect_snapshot({
     keyed_build_join_spec(x, y) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(x, y)) %>%
       dm_paste(options = c("select", "keys"))
     keyed_build_join_spec(y, x) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(y, x)) %>%
       dm_paste(options = c("select", "keys"))
   })
@@ -262,11 +264,11 @@ test_that("joins with same child PK", {
 
   expect_snapshot({
     keyed_build_join_spec(x, y) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(x, y)) %>%
       dm_paste(options = c("select", "keys"))
     keyed_build_join_spec(y, x) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(y, x)) %>%
       dm_paste(options = c("select", "keys"))
   })
@@ -286,11 +288,11 @@ test_that("joins with same child PK and same name", {
 
   expect_snapshot({
     keyed_build_join_spec(x, y) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(x, y)) %>%
       dm_paste(options = c("select", "keys"))
     keyed_build_join_spec(y, x) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, r = left_join(y, x)) %>%
       dm_paste(options = c("select", "keys"))
   })
@@ -312,11 +314,11 @@ test_that("joins with other FK from parent", {
 
   expect_snapshot({
     keyed_build_join_spec(x, y) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, z, r = left_join(x, y)) %>%
       dm_paste(options = c("select", "keys"))
     keyed_build_join_spec(y, x) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, z, r = left_join(y, x)) %>%
       dm_paste(options = c("select", "keys"))
   })
@@ -338,11 +340,11 @@ test_that("joins with other FK from parent and name conflict", {
 
   expect_snapshot({
     keyed_build_join_spec(x, y) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, z, r = left_join(x, y)) %>%
       dm_paste(options = c("select", "keys"))
     keyed_build_join_spec(y, x) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, z, r = left_join(y, x)) %>%
       dm_paste(options = c("select", "keys"))
   })
@@ -364,11 +366,11 @@ test_that("joins with other FK from child", {
 
   expect_snapshot({
     keyed_build_join_spec(x, y) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, z, r = left_join(x, y)) %>%
       dm_paste(options = c("select", "keys"))
     keyed_build_join_spec(y, x) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, z, r = left_join(y, x)) %>%
       dm_paste(options = c("select", "keys"))
   })
@@ -390,11 +392,11 @@ test_that("joins with other FK from child and name conflict", {
 
   expect_snapshot({
     keyed_build_join_spec(x, y) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, z, r = left_join(x, y)) %>%
       dm_paste(options = c("select", "keys"))
     keyed_build_join_spec(y, x) %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
     dm(x, y, z, r = left_join(y, x)) %>%
       dm_paste(options = c("select", "keys"))
   })
@@ -405,14 +407,14 @@ test_that("left join works as expected with keyed tables", {
 
   expect_snapshot({
     dm <- dm_nycflights13()
-    keyed_tbl_impl(dm, "weather") %>% left_join(keyed_tbl_impl(dm, "flights"))
+    keyed_tbl_impl(dm, "weather") %>% left_join(keyed_tbl_impl(dm, "flights"), multiple = "all")
   })
 
   # results should be similar to zooming
-  zd1 <- dm_zoom_to(dm, weather) %>% left_join(flights)
+  zd1 <- dm_zoom_to(dm, weather) %>% left_join(flights, multiple = "all")
   zd2 <- dm_zoom_to(dm, flights) %>% left_join(weather)
 
-  jd1 <- keyed_tbl_impl(dm, "weather") %>% left_join(keyed_tbl_impl(dm, "flights"))
+  jd1 <- keyed_tbl_impl(dm, "weather") %>% left_join(keyed_tbl_impl(dm, "flights"), multiple = "all")
   jd2 <- keyed_tbl_impl(dm, "flights") %>% left_join(keyed_tbl_impl(dm, "weather"))
 
   expect_equal(ncol(jd1), ncol(jd2))
@@ -515,7 +517,21 @@ test_that("pks_df_from_keys_info()", {
     dm %>%
       dm_get_keyed_tables_impl() %>%
       pks_df_from_keys_info() %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
+  })
+})
+
+test_that("uks_df_from_keys_info()", {
+  withr::local_seed(20220715)
+
+  dm <- dm_for_filter() %>%
+    dm_add_uk(tf_5, l)
+
+  expect_snapshot({
+    dm %>%
+      dm_get_keyed_tables_impl() %>%
+      uks_df_from_keys_info() %>%
+      to_snapshot_json()
   })
 })
 
@@ -528,7 +544,7 @@ test_that("fks_df_from_keys_info()", {
     dm %>%
       dm_get_keyed_tables_impl() %>%
       fks_df_from_keys_info() %>%
-      jsonlite::toJSON(pretty = TRUE)
+      to_snapshot_json()
   })
 })